+++ wled00/fcn_declare.h
@@ -267,7 +267,7 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply=tru
 
 //udp.cpp
 void notify(byte callMode, bool followUp=false);
-uint8_t realtimeBroadcast(uint8_t type, IPAddress client, uint32_t length, uint8_t* buffer, uint8_t bri = 255, bool isRGBW = false, uint8_t outouts = 1, uint16_t leds_per_output = 1, uint8_t fps_limit = 1, uint8_t color_order = 0, bool e131_multicast = false);
+uint8_t realtimeBroadcast(uint8_t type, IPAddress client, uint32_t length, uint8_t* buffer, uint8_t bri = 255, bool isRGBW = false, uint32_t outouts = 1, uint32_t leds_per_output = 1, uint8_t fps_limit = 1, uint8_t color_order = 0, bool e131_multicast = false);
 void realtimeLock(uint32_t timeoutMs, byte md = REALTIME_MODE_GENERIC);
 void exitRealtime();
 void handleNotifications();
@@ -396,7 +396,7 @@ void userLoop();
 
 //util.cpp
 void dumpAllTaskHWMs(void);
-int getNumVal(const String* req, uint16_t pos);
+int getNumVal(const String* req, uint32_t pos);
 void parseNumber(const char* str, byte* val, byte minv=0, byte maxv=255);
 bool getVal(JsonVariant elem, byte* val, byte minv=0, byte maxv=255);
 bool updateVal(const char* req, const char* key, byte* val, byte minv=0, byte maxv=255);
